Got the Maplin arm driver working a bit better Created 'ground truth' representations of what the detected motion blocks should be

+++ gaffa_apps/gripper_detector/src/gripper_detector/OpticalFlowFilter.py
@@ -22,11 +22,11 @@ class OpticalFlowFilter:
 
     #---------------------------------------------------------------------------
     def calcOpticalFlowWidth( self, imageWidth ):
-        return (imageWidth - self.opticalFlowBlockWidth)/self.opticalFlowBlockWidth
+        return int((imageWidth - self.opticalFlowBlockWidth)/self.opticalFlowBlockWidth)
         
     #---------------------------------------------------------------------------
     def calcOpticalFlowHeight( self, imageHeight ):
-        return (imageHeight - self.opticalFlowBlockHeight)/self.opticalFlowBlockHeight
+        return int((imageHeight - self.opticalFlowBlockHeight)/self.opticalFlowBlockHeight)
         
     #---------------------------------------------------------------------------
     def calcOpticalFlow( self, curImageGray ):
